+++ Raylib-cs/types/Mesh.cs
@@ -5,28 +5,6 @@ using System.Runtime.InteropServices;
 namespace Raylib_cs
 {
     /// <summary>
-    /// Transform, vectex transformation data
-    /// </summary>
-    [StructLayout(LayoutKind.Sequential)]
-    public partial struct Transform
-    {
-        /// <summary>
-        /// Translation
-        /// </summary>
-        public Vector3 translation;
-
-        /// <summary>
-        /// Rotation
-        /// </summary>
-        public Vector4 rotation;
-
-        /// <summary>
-        /// Scale
-        /// </summary>
-        public Vector3 scale;
-    }
-
-    /// <summary>
     /// Vertex data definning a mesh<br/>
     /// NOTE: Data stored in CPU memory (and GPU)
     /// </summary>
